requirements.py.txt
Jinja2==3.1.2
MarkupSafe==2.1.1
PyYAML==6.0
simplejson==3.17.6
